Rename binary to yopass-server and introduce pkg/server
In preparation of introducing a yopass CLI command this change frees up
the cmd name "yopass" by renaming the server to "yopass-server". Moving
the types and functions only relevant to the server to a new Go package
will allow to build a smaller Go CLI binary without having to include
all server dependencies.