Return if an item was deleted by the database
In order to know if an item was deleted or not found without implementing implementation based error checking a boolean should also be returned by Delete.

If an item is Deleted then the return will be `true, nil`, if an item is not found the return is `false, nil` and if an error correct the return is `false, error`.